Added image atlases for workers, flags, trees, animals, and houses. Image offsets for houses are incorrect in some places for other nations than romans. Also many fixes.

+++ src/main/java/org/appland/settlers/assets/Unsigned.java
@@ -68,7 +68,7 @@ public class Unsigned {
 
         for (int i = 0; i < 8; i++) {
             System.out.println ("" + i + ": "
-                    + Integer.toHexString ((int)getUnsignedByte (buffer, i)));
+                    + Integer.toHexString (getUnsignedByte (buffer, i)));
         }
 
         System.out.println ("2: " + Integer.toHexString (getUnsignedShort (buffer, 2)));
